@@ -1,22 +1,22 @@
 import test from 'ava';
+import chalk from '../source/index.js';
 
+// TODO: Do this when ESM supports loader hooks
 // Spoof supports-color
-require('./_supports-color')(__dirname, {
-	stdout: {
-		level: 0,
-		hasBasic: false,
-		has256: false,
-		has16m: false
-	},
-	stderr: {
-		level: 0,
-		hasBasic: false,
-		has256: false,
-		has16m: false
-	}
-});
-
-const chalk = require('../source');
+// require('./_supports-color')(__dirname, {
+// 	stdout: {
+// 		level: 0,
+// 		hasBasic: false,
+// 		has256: false,
+// 		has16m: false
+// 	},
+// 	stderr: {
+// 		level: 0,
+// 		hasBasic: false,
+// 		has256: false,
+// 		has16m: false
+// 	}
+// });
 
 test('colors can be forced by using chalk.level', t => {
 	chalk.level = 1;
